Restored UpdateFrequencyWrapper and ScalingWrapper, changed isinstance
isinstance checks on component base types now check that the instance
satisfies properties of that base type rather than strict type checking.
This allows a single wrapper type to have instances that can be any
of a number of component types, depending on the wrapped component type.

UpdateFrequencyWrapper is again a wrapper, which can now wrap any
component type. Similarly for ScalingWrapper. Documentation still has to
be updated for these wrappers.