Fixed bug in update_dict_by_adding_another
Bug would take place when initial dictionary had a DataArray which
does not have a dimension in the second dictionary. In that case
in-place addition cannot take place because the dimensions of the
first array are not mutable. Fixed by replacing the array with a
new array (the dict is still updated in-place with the new array).